requirements.txt
flask
flask-sqlalchemy
flask-cors
pandas
marshmallow
python-dotenv
psycopg2-binary
flask-jwt-extended
werkzeug
